full schedule

+++ lib/screens/qpi/add_qpi.dart
@@ -3,17 +3,18 @@ import 'package:provider/provider.dart';
 
 import 'package:admu_student_app/constants/app_colors.dart';
 import 'package:admu_student_app/models/academic_records.dart';
-import 'package:admu_student_app/models/add_qpi_notifier.dart';
 import 'package:admu_student_app/models/course.dart';
 import 'package:admu_student_app/models/semester.dart';
 import 'package:admu_student_app/models/year.dart';
 import 'package:admu_student_app/screens/qpi/add_course.dart';
 import 'package:admu_student_app/screens/qpi/add_semester.dart';
 import 'package:admu_student_app/screens/qpi/add_year.dart';
+import 'package:admu_student_app/widgets/modals/alert.dart';
+import 'package:admu_student_app/widgets/modals/custom_snack_bar.dart';
 import 'package:admu_student_app/widgets/button_row.dart';
 import 'package:admu_student_app/widgets/buttons.dart';
 
-class AddQPIPage extends StatelessWidget {
+class AddQPIPage extends StatefulWidget {
   final int yearNum;
   final int semNum;
 
@@ -24,9 +25,6 @@ class AddQPIPage extends StatelessWidget {
   final int units;
 
   final bool isEditing;
-  // 0 = year
-  // 1 = sem
-  // 2 = course
   final int selected;
 
   AddQPIPage({
@@ -41,68 +39,24 @@ class AddQPIPage extends StatelessWidget {
   });
 
   @override
-  Widget build(BuildContext context) {
-    AddQPINotifier provider;
-
-    if (course != null)
-      provider = AddQPINotifier.fromCourse(semNum, course);
-    else
-      provider = AddQPINotifier.fromYearSem(
-          semester != null ? semester.semNum : semNum);
-
-    return ChangeNotifierProvider(
-      create: (_) => provider,
-      child: _AddQPI(
-        yearNum: yearNum,
-        semNum: semNum,
-        year: year,
-        semester: semester,
-        course: course,
-        units: units,
-        isEditing: isEditing,
-        selected: selected,
-      ),
-    );
-  }
+  _AddQPIPageState createState() => _AddQPIPageState();
 }
 
-class _AddQPI extends StatefulWidget {
-  final int yearNum;
-  final int semNum;
-
-  final Year year;
-  final Semester semester;
-  final Course course;
-
-  final int units;
-
-  final bool isEditing;
-  final int selected;
-
-  _AddQPI({
-    this.yearNum,
-    this.semNum,
-    this.year,
-    this.semester,
-    this.course,
-    this.units,
-    this.isEditing,
-    this.selected,
-  });
-
-  @override
-  _AddQPIState createState() => _AddQPIState();
-}
-
-class _AddQPIState extends State<_AddQPI> {
+class _AddQPIPageState extends State<AddQPIPage> {
   TextEditingController _yearCtrl = TextEditingController();
   TextEditingController _qpiCtrl = TextEditingController();
   TextEditingController _unitsCtrl = TextEditingController();
   TextEditingController _codeCtrl = TextEditingController();
 
+  int _semNum = 1;
+  Color _courseColor;
+  int _gradeVal = 1;
+
   int selected = 0;
   final List<String> titles = ['Year', 'Semester', 'Course'];
   List<Widget> screens = [];
+
+  final List<double> GRADES = [4.0, 3.5, 3.0, 2.5, 2.0, 1.0, 0.0];
 
   @override
   void initState() {
@@ -114,12 +68,25 @@ class _AddQPIState extends State<_AddQPI> {
     if (widget.year != null) {
       _unitsCtrl.text = '${widget.year.units}';
     }
+    if (widget.semNum != null) {
+      _semNum = widget.semNum;
+    }
     if (widget.semester != null) {
       _unitsCtrl.text = '${widget.semester.units}';
+      _semNum = widget.semester.semNum;
     }
     if (widget.course != null) {
       _unitsCtrl.text = '${widget.course.units}';
       _codeCtrl.text = widget.course.courseCode;
+
+      _courseColor = widget.course.color;
+
+      for (int i = 0; i < GRADES.length; i++) {
+        if (GRADES[i] == widget.course.qpi) {
+          _gradeVal = i + 1;
+          break;
+        }
+      }
     }
     if (widget.isEditing && widget.selected == 0) {
       _yearCtrl.text = '${widget.year.yearNum}';
@@ -138,20 +105,44 @@ class _AddQPIState extends State<_AddQPI> {
         yearController: _yearCtrl,
         unitsController: _unitsCtrl,
         qpiController: _qpiCtrl,
+        selected: _semNum,
+        onValueChange: _onSemChange,
       ),
       CourseAddQPI(
         yearController: _yearCtrl,
         unitsController: _unitsCtrl,
         codeController: _codeCtrl,
+        color: _courseColor,
+        onColorChange: _onColorChange,
+        sem: _semNum,
+        onSemChange: _onSemChange,
+        gradeVal: _gradeVal,
+        onGradeChange: _onGradeChange,
       ),
     ];
 
     selected = widget.selected > 2 ? 0 : widget.selected;
   }
 
-  void _onSave(int semNum, int gradeVal) async {
-    const List<double> GRADES = [4.0, 3.5, 3.0, 2.5, 2.0, 1.0, 0.0];
+  void _onColorChange(Color color) {
+    setState(() {
+      _courseColor = color;
+    });
+  }
 
+  void _onSemChange(int val) {
+    setState(() {
+      _semNum = val;
+    });
+  }
+
+  void _onGradeChange(int val) {
+    setState(() {
+      _gradeVal = val;
+    });
+  }
+
+  void _onSave() async {
     // no error handling
     int yearNum = int.parse(_yearCtrl.text);
     int units = int.parse(_unitsCtrl.text);
@@ -159,85 +150,113 @@ class _AddQPIState extends State<_AddQPI> {
 
     if (selected == 0) {
       double qpi = double.parse(_qpiCtrl.text);
-      if (widget.isEditing)
+      if (widget.isEditing) {
         Provider.of<AcademicRecords>(context, listen: false).editYearlyQPI(
           widget.year,
           yearNum,
           units,
           qpi,
         );
-      else
+
+        CustomSnackBar.showSnackBar(context, 'Year QPI edited!');
+      } else {
         Provider.of<AcademicRecords>(context, listen: false).addYearlyQPI(
           yearNum,
           units,
           qpi,
         );
+
+        CustomSnackBar.showSnackBar(context, 'Year QPI added!');
+      }
     } else if (selected == 1) {
       double qpi = double.parse(_qpiCtrl.text);
-      if (widget.isEditing)
+      if (widget.isEditing) {
         Provider.of<AcademicRecords>(context, listen: false).editSemestralQPI(
           widget.yearNum,
           widget.semester,
           yearNum,
-          semNum,
+          _semNum,
           units,
           qpi,
         );
-      else
+
+        CustomSnackBar.showSnackBar(context, 'Semester QPI edited!');
+      } else {
         Provider.of<AcademicRecords>(context, listen: false).addSemestralQPI(
           yearNum,
-          semNum,
+          _semNum,
           units,
           qpi,
         );
+
+        CustomSnackBar.showSnackBar(context, 'Semester QPI added!');
+      }
     } else if (selected == 2) {
-      if (widget.isEditing)
+      if (widget.isEditing) {
         Provider.of<AcademicRecords>(context, listen: false).editCourse(
           widget.yearNum,
           widget.semNum,
           widget.course,
           yearNum,
-          semNum,
+          _semNum,
           code,
-          Provider.of<AddQPINotifier>(context, listen: false).color.value,
+          _courseColor.value,
           units,
-          GRADES[gradeVal - 1],
+          GRADES[_gradeVal - 1],
           true, // no handling
         );
-      else
+
+        CustomSnackBar.showSnackBar(context, 'Class QPI edited!');
+      } else {
         Provider.of<AcademicRecords>(context, listen: false).addCourse(
           yearNum,
-          semNum,
+          _semNum,
           code,
-          Provider.of<AddQPINotifier>(context, listen: false).color.value,
+          _courseColor.value,
           units,
-          GRADES[gradeVal - 1],
+          GRADES[_gradeVal - 1],
           true, // no handling for non-included
         );
+
+        CustomSnackBar.showSnackBar(context, 'Class QPI added!');
+      }
     }
 
     Navigator.of(context).pop();
   }
 
   void _onDelete() async {
-    if (selected == 0)
+    bool toDelete = false;
+
+    await AlertModal.showAlert(
+      context,
+      header: 'Delete QPI',
+      onAccept: () {
+        toDelete = true;
+      },
+    );
+
+    if (!toDelete) return;
+
+    if (selected == 0) {
       Provider.of<AcademicRecords>(context, listen: false)
           .deleteYearlyQPI(widget.yearNum);
-    else if (selected == 1)
+      CustomSnackBar.showSnackBar(context, 'Year QPI deleted!');
+    } else if (selected == 1) {
       Provider.of<AcademicRecords>(context, listen: false)
           .deleteSemestralQPI(widget.yearNum, widget.semester.semNum);
-    else if (selected == 2)
+      CustomSnackBar.showSnackBar(context, 'Semester QPI deleted!');
+    } else if (selected == 2) {
       Provider.of<AcademicRecords>(context, listen: false).deleteCourse(
           widget.yearNum, widget.semNum, widget.course.courseCode);
+      CustomSnackBar.showSnackBar(context, 'Class QPI deleted!');
+    }
 
     Navigator.of(context).pop();
   }
 
   @override
   Widget build(BuildContext context) {
-    int sem = Provider.of<AddQPINotifier>(context).semNum;
-    int gradeVal = Provider.of<AddQPINotifier>(context).gradeVal;
-
     return Scaffold(
       resizeToAvoidBottomInset: false,
       appBar: AppBar(
@@ -246,7 +265,7 @@ class _AddQPIState extends State<_AddQPI> {
           Padding(
             padding: EdgeInsets.fromLTRB(0, 0, 16, 0),
             child: TextButton(
-              onPressed: () => _onSave(sem, gradeVal),
+              onPressed: _onSave,
               child: Text('Done',
                   style: Theme.of(context)
                       .textTheme
@@ -256,10 +275,9 @@ class _AddQPIState extends State<_AddQPI> {
           ),
         ],
         title: IconButton(
-            icon: Icon(Icons.close_rounded),
-            onPressed: () {
-              Navigator.pop(context);
-            }),
+          icon: Icon(Icons.close_rounded),
+          onPressed: () => Navigator.pop(context),
+        ),
         elevation: 0,
       ),
       backgroundColor: AppColors.PRIMARY_MAIN,
@@ -268,20 +286,14 @@ class _AddQPIState extends State<_AddQPI> {
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.center,
           children: [
-            Container(
-              padding: EdgeInsets.fromLTRB(0, 0, 0, 40),
-              child: Row(
-                children: [
-                  Text(
-                    '${widget.isEditing ? 'Edit' : 'Add'} ${titles[selected]} QPI',
-                    style: Theme.of(context)
-                        .textTheme
-                        .headline4
-                        .copyWith(color: AppColors.GRAY_LIGHT[2]),
-                  )
-                ],
-              ),
+            Text(
+              '${widget.isEditing ? 'Edit' : 'Add'} ${titles[selected]} QPI',
+              style: Theme.of(context)
+                  .textTheme
+                  .headline4
+                  .copyWith(color: AppColors.GRAY_LIGHT[2]),
             ),
+            SizedBox(height: 40),
             ButtonRow('Year', 'Semester', 'Course', () {
               // Item 1 Clicked
               setState(() {
@@ -300,10 +312,14 @@ class _AddQPIState extends State<_AddQPI> {
             }, selected, widget.isEditing),
             SizedBox(height: 24),
             screens[selected],
-            // Spacer(),
-            SizedBox(height: 96),
+            SizedBox(height: 56),
             widget.isEditing
-                ? LongButton('Delete', Colors.orange, Colors.white, _onDelete)
+                ? LongButton(
+                    'Delete',
+                    AppColors.SECONDARY_MAIN,
+                    AppColors.GRAY_LIGHT[2],
+                    _onDelete,
+                  )
                 : Container(),
           ],
         ),
